Introduce the CustomizeAttribute abstract class in the AutoFixture library
Obsolete old CustomizeAttribute class in the xUnit.net project
