+++ Src/AutoFixture.xUnit.net/CustomizeAttributeComparer.cs
@@ -2,9 +2,9 @@
 
 namespace Ploeh.AutoFixture.Xunit
 {
-    internal class CustomizeAttributeComparer : Comparer<CustomizeAttribute>
+    internal class CustomizeAttributeComparer : Comparer<AutoFixture.CustomizeAttribute>
     {
-        public override int Compare(CustomizeAttribute x, CustomizeAttribute y)
+        public override int Compare(AutoFixture.CustomizeAttribute x, AutoFixture.CustomizeAttribute y)
         {
             var xfrozen = x is FrozenAttribute;
             var yfrozen = y is FrozenAttribute;
